Updates to make this flag-based instead of environment based, also moved logic to config section
