+++ mongo/mongo.go
@@ -2,21 +2,17 @@ package mongo
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net"
 	"os"
 	"reflect"
-	"regexp"
 	"strings"
 	"sync"
 	"time"
 	"unsafe"
 
 	"github.com/DataDog/datadog-go/statsd"
-	"github.com/coinbase/mongobetween/util"
-	"go.mongodb.org/mongo-driver/event"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/address"
 	"go.mongodb.org/mongo-driver/mongo/description"
@@ -62,12 +58,6 @@ func Connect(log *zap.Logger, sd *statsd.Client, opts *options.ClientOptions, pi
 	ctx, cancel := context.WithTimeout(context.Background(), pingTimeout)
 	defer cancel()
 
-	opts = opts.SetPoolMonitor(poolMonitor(sd))
-
-	if envVarIsTrue("ENABLE_SERVER_MONITORING") {
-		opts = opts.SetServerMonitor(serverMonitoring(log, sd))
-	}
-
 	var err error
 	log.Info("Connect")
 	c, err := mongo.Connect(ctx, opts)
@@ -102,117 +92,6 @@ func Connect(log *zap.Logger, sd *statsd.Client, opts *options.ClientOptions, pi
 	go m.cacheMonitor()
 
 	return &m, nil
-}
-
-func poolMonitor(sd *statsd.Client) *event.PoolMonitor {
-	checkedOut, checkedIn := util.StatsdBackgroundGauge(sd, "pool.checked_out_connections", []string{})
-	opened, closed := util.StatsdBackgroundGauge(sd, "pool.open_connections", []string{})
-
-	return &event.PoolMonitor{
-		Event: func(e *event.PoolEvent) {
-			snake := strings.ToLower(regexp.MustCompile("([a-z0-9])([A-Z])").ReplaceAllString(e.Type, "${1}_${2}"))
-			name := fmt.Sprintf("pool_event.%s", snake)
-			tags := []string{
-				fmt.Sprintf("address:%s", e.Address),
-				fmt.Sprintf("reason:%s", e.Reason),
-			}
-			switch e.Type {
-			case event.ConnectionCreated:
-				opened(name, tags)
-			case event.ConnectionClosed:
-				closed(name, tags)
-			case event.GetSucceeded:
-				checkedOut(name, tags)
-			case event.ConnectionReturned:
-				checkedIn(name, tags)
-			default:
-				_ = sd.Incr(name, tags, 1)
-			}
-		},
-	}
-}
-
-func serverMonitoring(log *zap.Logger, statsdClient *statsd.Client) *event.ServerMonitor {
-
-	return &event.ServerMonitor{
-
-		ServerOpening: func(e *event.ServerOpeningEvent) {
-			_ = statsdClient.Incr("server_opening_event",
-				[]string{
-					fmt.Sprintf("address:%s", e.Address),
-					fmt.Sprintf("topology_id:%s", e.TopologyID.Hex()),
-				}, 0)
-		},
-
-		ServerClosed: func(e *event.ServerClosedEvent) {
-			_ = statsdClient.Incr("server_closed_event",
-				[]string{
-					fmt.Sprintf("address:%s", e.Address),
-					fmt.Sprintf("topology_id:%s", e.TopologyID.Hex()),
-				}, 0)
-		},
-
-		ServerDescriptionChanged: func(e *event.ServerDescriptionChangedEvent) {
-			_ = statsdClient.Incr("server_description_changed_event",
-				[]string{
-					fmt.Sprintf("address:%s", e.Address),
-					fmt.Sprintf("topology_id:%s", e.TopologyID.Hex()),
-				}, 0)
-
-			if envVarIsTrue("ENABLE_SERVER_MONITORING_LOGGING") {
-				var prevDMap map[string]interface{}
-				var newDMap map[string]interface{}
-
-				prevDescription, _ := json.Marshal(&e.PreviousDescription)
-				_ = json.Unmarshal(prevDescription, &prevDMap)
-				newDescription, _ := json.Marshal(e.NewDescription)
-				_ = json.Unmarshal(newDescription, &newDMap)
-
-				log.Info("ServerDescriptionChangedEvent detected. ",
-					zap.Any("address", e.Address),
-					zap.String("topologyId", e.TopologyID.Hex()),
-					zap.Any("prevDescription", prevDMap),
-					zap.Any("newDescription", newDMap),
-				)
-			}
-		},
-
-		TopologyDescriptionChanged: func(e *event.TopologyDescriptionChangedEvent) {
-			_ = statsdClient.Incr("topology_description_changed_event",
-				[]string{
-					fmt.Sprintf("topology_id:%s", e.TopologyID.Hex()),
-				}, 0)
-			if envVarIsTrue("ENABLE_SERVER_MONITORING_LOGGING") {
-				var prevDMap map[string]interface{}
-				var newDMap map[string]interface{}
-
-				prevDescription, _ := json.Marshal(&e.PreviousDescription)
-				_ = json.Unmarshal(prevDescription, &prevDMap)
-				newDescription, _ := json.Marshal(e.NewDescription)
-				_ = json.Unmarshal(newDescription, &newDMap)
-
-				log.Info("TopologyDescriptionChangedEvent detected. ",
-					zap.String("topologyId", e.TopologyID.Hex()),
-					zap.Any("prevDescription", prevDMap),
-					zap.Any("newDescription", newDMap),
-				)
-			}
-		},
-
-		TopologyOpening: func(e *event.TopologyOpeningEvent) {
-			_ = statsdClient.Incr("topology_opening_event",
-				[]string{
-					fmt.Sprintf("topology_id:%s", e.TopologyID.Hex()),
-				}, 0)
-		},
-
-		TopologyClosed: func(e *event.TopologyClosedEvent) {
-			_ = statsdClient.Incr("topology_closed_event",
-				[]string{
-					fmt.Sprintf("topology_id:%s", e.TopologyID.Hex()),
-				}, 0)
-		},
-	}
 }
 
 func (m *Mongo) Description() description.Topology {
